requirements.txt
asgiref==3.6.0
Django==4.1.7
django-cors-headers==3.13.0
django-rest-framework==0.1.0
djangorestframework==3.14.0